pidl: Do not produce a function that is replaced by MANUAL code
If a MANUAL directive in a conformance file causes a function not to be
emitted, do not bother calculating the function that will not be used.
This is similar to the NOEMIT directive, but has different logic because
MANUAL only prevents emitting the function for one level of an element
instead of all the functions and variables. This does not change the
dissectors produced at all, only skips some unnecessary compilation.

In particular, the messages when compiling the pidl-dissectors target:

dnsserver.idl:159: error: Inline arrays not supported
eventlog.idl:54: error: Inline arrays not supported

are no longer inaccurately produced. The inline arrays in those two IDL
files have long been supported through the MANUAL code blocks.

Expand on the error message to suggest the use of MANUAL directives
for implementation.

Remove the comments in the README and CMakeLists.txt about those error
messages being produced.