TFTP: Fix conversation lookups
Both find_conversation_strat calls are using the same direction; we
need to check both (we don't know which side is the original client
port).

Always create a new conversation for a read or write request (which
is always the first TFTP opcode according to RFC 1350.) The
tftp_conv_info_t struct only handles data for a single transfer;
we can't reuse it for another transfer and have things work on the
second pass. (Maybe the conversation data should be a wmem_tree_t
of these?)

Do so even if the port doesn't match, as we could have gotten here
from, e.g., a conversation dissector being sent (if ports are reused
in the same capture), in which case match_uint might be 17 for UDP
from the IP dissector, say.

Fixes the handling of the files in #16767 again.

Fixup a37b6c166929141a9044abe1520f13bc4894a893


(cherry picked from commit 401a10460d07f280e4ad797092c05fa4d6338767)

Co-authored-by: John Thacker <johnthacker@gmail.com>