wkssvc: Fix platformID enum values
pidl does not properly generate dissectors for typedefs where the
type specifier returns to another user defined type, instead of
a base type or a struct, union, or enum directly in the typedef
statement.

Use the conformance file to refer to the publicly exported
srvsvc enum values. Replace the HF_FIELD command (which wasn't
being used) with a TYPE command to do so.