TDS: Prevent signed overflow
Don't fetch the featureackdatalen before we know if it's present
(i.e., if the featureid is the terminator 0xff) but instead get it
with proto_tree_add_item_ret_uint. Set the length of the parent tree
after adding the feature data, which will throw an exception if
overflow would occur.

OSS-Fuzz 450116221