Always use the report_message routines.
Instead of using failure_message and alert_box routines, use the report_
routines that they end up calling. Make the failure_message and
alert_box routines static.

This makes code that reports various errors and warnings more
consistent, and means that the internal behavior of the report_
routines, and the behavior of the routines it calls, can be changed, as
long as the exernal behavior of the report_ routines remains compatible.