packaging: Windows version checks
While Qt has officially only supported Windows 10 1809 or later since
Qt 6.2[1], in practice our releases with Qt 6.5 have run on Windows
10 version 1607 and Windows Server 2016 (but not Windows 10 1507[2]).
However, Qt 6.8 and later doesn't run on those versions, as qt6core.dll
calls SetThreadDescription, which isn't available on Windows 10 version 1607
and Windows Server 2016 without Run Time Dynamic Linking that Qt doesn't
do by default.[3,4] It is possible to compile Qt 6.9.1 and later with
a newly added flag to re-enable Windows Server 2016 (and Windows 10
version 1607) support, but I'm not sure we want to do that even on 4.6;
certainly not for the 5.0 release.[5]

Because some of the Windows version variables were frozen with the
transition to Windows as a Service, and with the difference in the
Win32 API available at various build versions (all called Windows
10 and with the same VersionNT[6]), we have to check the build number
instead.

Update NSIS, WIX, and various other places for the minimum of
Windows 10 1809, aka Redstone 5 aka build number 17763. (Which
also corresponds to Windows Server 2019.)

[1] - https://doc.qt.io/archives/qt-6.2/supported-platforms.html
[2] - https://ask.wireshark.org/question/37410/qt6-entrypoint-could-not-be-found/
[3] - https://ask.wireshark.org/question/37552/wireshark-460-qt6coredll-setthreaddescription-error-on-server-2016/
[4] - https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-setthreaddescription
[5] - https://bugreports.qt.io/browse/QTBUG-134075
[6] - https://learn.microsoft.com/en-us/troubleshoot/windows-client/application-management/versionnt-value-for-windows-10-server