expert: Track expert info with the same abbrev
It's possible to register multiple expert info with the same abbrev.
In order for freeing and deregistering to work properly, they have
to be tracked somehow. (This apparently didn't cause a problem before
because nothing was actually freed or had something that needed to
be set to NULL to avoid trying to free a static string.)

Fixup 0e64c8871d48c3f8dfbe0a92561c7fe002f2031b
Fix #20794