NTLMSSP: Check for negative offsets
NTLMSSP has a unsigned 32 bit number that is used to mean an offset.
tvbs have signed 32 bit offsets; negative numbers mean "backwards from
the end of the tvb" which isn't what this means. Throw an exception
if the retrieved offset is signed, to avoid undefined behavior,
adding the wrong bytes as a string, and possibly even moving backwards.

OSS-Fuzz 451333962