Qt: Hide the DataSourceTab before adding or removing tabs
When a tab is added or removed, the QTabBar can re-layout all the tabs,
which calculates the sizeHint and minimumSizeHint for every tab. If
it is called every time a tab is added or removed, it becomes an O(N^2)
operation to add or remove N tabs.

Making the DataSourceTab not visible before adding or removing tabs
and then visible again at the end can in some cases remove this to
a single layoutTabs call, reducing things to O(N).

Unfortunately a Qt commit in 6.3.0 means that layoutTabs is always
called when inserting a new tab, even if the QTabWidget is hidden:
https://bugreports.qt.io/browse/QTBUG-141187

However, on all versions of Qt this makes removing when switching to
a new frame much faster, so this addresses part of #20797


(cherry picked from commit 2a09c1a676ee10ae6bcb9a9beab749b5e15a5636)

Co-authored-by: John Thacker <johnthacker@gmail.com>