Qt: Put back the clear call in DataSourceTab
It's not needed on Linux, but apparently is on macOS and maybe Windows.
On my Linux and Windows tests with Qt 6.9 the delete operation is still O(N)
instead of O(N^2) even with the clear call; hopefully that's true
in general.

Fixup 2a09c1a676ee10ae6bcb9a9beab749b5e15a5636