wsutil: Use a common struct for SI/IEC prefixes.
Consolidate the logic for the adjustments needed between SI and IEC prefixes
into one shared data structure ("const"ed, of course) and use it across
format_units and format_size_wmem.  As a result, format_size_wmem will now use
the "peta-" prefix too, but we lose the manually-unrolled loop in that
function.  (The "exa-" prefix won't be used until we support 128-bit integer
data types...)

Shuffle around the structure of the scaling loops in format_units, hopefully in
a way that makes static analyzers more happy that we won't run off the end of
the prefix array.

Add test coverage for the "is_small" case and the upper limits in
format_size_wmem.

Zap some #defines which have been unused for several years.