extcap: Do not try to read c:\etc\ssh\ssh_config on Windows
libssh by default tries to read /etc/ssh/ssh_config on all OSes.
OpenSSH Portable for Windows by default doesn't create a systemwide
ssh_config, and if it does, it puts it in %ProgramData%\ssh and
has to check permissions because it might be owned and writable
by a non-admin user. Since C:\etc does not exist by default on
Windows, that's a possible security issue.

https://github.com/PowerShell/openssh-portable/blob/latestw_all/ssh.c#L591-L597
https://github.com/PowerShell/openssh-portable/blob/latestw_all/contrib/win32/openssh/config.h.vs#L1723

Only read from the OpenSSH default configuration file in the user
home directory on Windows.

Fix #20806.