pidl-wireshark: Handle unions with identical cases
Handle unions where multiple case values are the same and share
a name, without having to rename them. We need to output the
code calls in the switch statement without outputting the
function declarations and implementations a second time.

We don't have any idl files like this in the Wireshark repository,
but some in the Samba repository or that Microsoft publishes that
we don't use (yet) have this.