PER: Don't add data sources that aren't new
When dissecting a ASN.1 BIT STRING in PER, if the bit string is
both byte aligned and an even number of bytes, then the aligned
tvb produced is merely a subset tvb of the one passed in and does
not have a new data source tvb. In such a case, don't add it as
a new data source.

Adding a new data source for a tvb that shares ds_tvb with another
data source is not helpful in the GUI, as clicking on the field in
the proto tree will select the earlier tab with the same ds_tvb.
(Clicking on the bytes in the extra tab also does not appear to work.)

On frame 16 in the capture in #20797 this reduces the number of data
source tabs from 4594 to 4027, as there are a large number of 16 bit
BIT STRINGs.