Qt: Workaround QTabBar O(N^2) bug
Workaround https://bugreports.qt.io/browse/QTBUG-141187 on
Qt 6.3 and higher.

Create a class derived from QTabBar to short circuit tabSizeHint
and minimumTabSizeHint when the QTabBar is not visible. This isn't
the perfect solution, as the functions that really need to be
short-circuited aren't virtual and can't be overridden unlike these
two. However, while adding lots of tabs is still O(N^2) due to
all the layoutTabs() calls, it is about ten times faster and for
the case in #20797 reduces the time to select the packet from a
close to unusable 20 seconds to a merely painful 2 s.


(cherry picked from commit 362dd5aa9fbe05a82d80d5f46748e8b270290e10)

Co-authored-by: John Thacker <johnthacker@gmail.com>