wsutil: Pull out "application flavor" to its own folder
Pull out the functionality of "application flavor" and put it into a separate folder, one for Wireshark (packets) and one for Stratoshark (events).  Then applications only have to link the appropriate module that they support, either packets or events.  There's no need for an explicit setting of behavior, just link to the desired module. This works towards better isolating the differences between Wireshark and Stratoshark and not having explicit checks for which application is running.