writecap: Fix linkage on MSYS2
MSYS2 has difficulty figuring out that it needs to link in
ws_cwstream_write when using the static libwritecap after
commit 40b552fee6cc9f2ed11d42ed43ca313343d7c9d8, even when
linking udpdump against wustil. Probably some kind of
dllimport nonsense. (It doesn't have a problem with dumpcap,
which links against the static wsutil.)

https://github.com/wireshark/wireshark/actions/runs/18686056163/job/53278883781

This doesn't change the output libraries or binaries on
Linux or with MSVC.