CMake: Prefer WIRESHARK_BASE_DIR and make it more convenient
The CI runners used to derive the location for third party libraries
based on the WIRESHARK_LIB_DIR environment variable, but that hasn't
been the case for many years. Deprecate it in the Developer's Guide and
in CMake, and have WIRESHARK_BASE_DIR take precedence.

Allow WIRESHARK_BASE_DIR to be set using an environment variable or a
CMake variable on Windows and macOS.