Qt: Hide Go First and Last conversation packet menu items
With the loading and closing of a capture file the go to packet menu
items become enabled and disabled respectively. This should also be
true for the Go First conversation packet and Go Last conversation
packet. This change adds these operations.
It also matches the sort of the operations to the menu item order.