X11: Check for overflow
There are many "list of X" handlers where a parameter, which in some
cases can be a signed or unsigned 32 bit integer, gives the number of
X items present. Multiplying this by the size of X to get the total
length of a list can cause overflow. The number of items should never
be negative.

Make the parameter unsigned and check for overflow when multiplying
to prevent undefined behavior.
All the list type fields are FT_NONE, so this will make them be
added to the tree to the end of the tvb. An exception will be thrown
trying to add an individual item that runs off the tvb.

Also, in two places don't use the macros designed for when the value
obtained is the number of bytes, because that just multiplies and then
divides by the same constant 2 or 4 (but can overflow.)

OSS-Fuzz Issue 452385937