WSDG: Update the "Contribute your changes" page.
A few minor cleanups:

Add numbers to the Git triangular workflow diagram so that it's easier for
newbies to follow.  (I remember being quite confused when I first saw it and
was still learning git!)

Update the instructions for building the SVG diagram from the Graphviz source
to reflect the new doc directory layout.

Update the branch name for the example to something recent.

Add a bit more color to the description for commit messages.

Give guidance for branch selection when creating the MR for cherry-picks since
it's also something that newbies seem to trip over.

Zap some spurious leading '$' from the output of "git remote -v".