Qt: Display IO Graphs and Plots that fail validation
The IO Graphs and Plots validate that Y fields are valid registered
fields. Sometimes field abbreviations change between releases.

uat_load returns failure and an error if a record fails validation,
but still can add it in that case. It makes more sense to display
the record which failed validation in the GUI so that the user can
try to fix it rather than clear all records.

Fix #20788