wsutil: Fix NSTIME_INIT_MAX initializer
On many platforms (e.g. Windows), LONG_MAX and INT_MAX are the same,
and both are 32-bit even when time_t is 64-bit.

Move the more general TIME_T_MIN and TIME_T_MAX macros (which also
handle unsigned integer time_t) to the header and use those to set
the maximum possible value in the initializer.