wiretap: add a separate public header file for wiretap modules.
For built-in and plugin modules that handle file formats, have a
wtap_module.h header that declares what they need, and move everything
intended for use by modules from wtap-int.h to that header.

Make it public, for the benefit of third-party plugin modules.

The only stuff remaining in wtap-int.h is for use by code internal to
Wireshark.

Do some "include what you use" - and some "don't include what you don't
use" - work, and update a long-out-of-date comment,  whie we're at it.