NFS: Fix detection of NFS4 packets
NFS4 uses port 2049, it is RPC service, but does not use portmapper. So
port detection based on RPC portmapper does not work for this RPC service
because there is no portmapper traffic.

And if traffic is between some SSL service port (e.g. 993) and NFS4 port
2049 then the packet is decoded as SSL. Note that for NFS traffic it common
that client port is below 1024, and so there is a high chance to choose
some SSL service port.

Fix this problem and explicitly register also TCP port 2049 for RPC
dissector, like it is already done for TCP port 111 (used by RCP portmapper
service).

With this change TCP traffic between ports 993 and 2049 is decoded as NFS4.