NFS: Add support for NFSv4 filename snooping
One RPC NFS4 procedure call can contain (and in most cases always contains)
more operations which changes filehandle between operations. So for correct
NFSv4 filename snooping it is required to track not only RPC packet number
(xid) but also index in the list of operations calls in that RPC packet.

In NFSv4, only replies for GETFH, GETATTR and READDIR operations might
return filename mapping. So extend all functions to take operation code and
index in the list of operations and track which operation was doing mapping.

Note that filename snooping for READDIR replies is not implemented yet. It
requires to match directory name from READDIR request with child name from
particular entry of READIR response together with filehandle attribute
attached to the entry of READDIR response. Which is slightly complicated.

This change implements only filename snooping for GETFH and GETATTR.