MIDI: Create a shared MIDI SysEx ID database.
Consolidate the divergent copies of the MIDI SysEx (System Exclusive Message)
identifier list from the rtp-midi dissector and the sysex dissector, the latter
of which is used to decode USB MIDI messages.

Add a script which collects the latest data from a webpage on midi.org and
updates the relevant source files: packet-midi-sysex-id.c, which has a
value_string table mapping each short and long ID to its corresponding label,
and packet-midi-sysex-id.h, which has a #define for each identifier's numerical
value.

Update the rtp-midi and sysex dissectors to use the central copy instead of
their own ones.  This de-duplication reduces libwireshark's read-only data size
by 15 kBytes.

Use the term "SysEx ID" instead of "Manufacturer ID" throughout the new code
because manufacturer IDs are a subset of SysEx IDs; I have not updated the
usages of "Manufacturer ID" in existing dissector code.

The list committed in this change is current as of 2025-11-02.