VLAN: Restore the dissection depth when leaving the dissector
The VLAN dissection depth should only count nested tags, it doesn't
need to keep increasing when a frame contains multiple VLAN tags that
are not nested, whether from "View as File Format", or from protocols
that can contain multiple Ethernet frames in a large frame, like
TECMP.

Remove the special workaround for TECMP (#16685) as it's not needed
anymore. That's a little cleaner because the TECMP dissector doesn't
have to know about the VLAN dissector.

Fix #20831