TCP: Handle a reassembly edge case
For HTTP on TLS, it's possible that the HTTP have no Content-Length,
so REASSEMBLE_AT_FIN gets set but also all bytes are consumed. In
that case, we still create a new msp, but we need to indicate that
we haven't seen the first segment for the way that the non out of
order reassembly works.

Fix #20835