ZigBee Direct: Don't use a packed struct
zbee_sec_ccm_decrypt() expects the nonce to be a array of
uint8_t of size ZBEE_SEC_CONST_NONCE_LEN. Just use that, as
packet-zbee-security and packet-zbee-nwk-gp do, instead of
creating a packed struct whose members are never really used.
(Among other things, the "frame_counter" member of that struct
is unnecessarily a uint32_t, as it needs to be copied in the
same little-endian order as in the back, and thus trying to
access the struct member as a uint32_t would require byte swapping
on Big-Endian systems anyway.)

The same dissection works on the sample files attached to !12052.

Fix #20776


(cherry picked from commit 81f95ab1fd2ef70ea82d7bcd781d044de5adb09e)

Co-authored-by: John Thacker <johnthacker@gmail.com>