BER: prevent possible signed overlow undefined behavior
In a ASN.1 CHOICE, it's possible for the CHOICE length to be
larger than that of the chosen child item. dissect_ber_choice
doesn't dissect the leftover bytes in that case, but always
returns the end offset based on the outer length.

That can overflow in the case where the tvb passed into
dissect_ber_choice is a subset of some other tvb, and it's added
later, which happens in SPNEGO and other protocols that call
Kerberos. To be safe, call ensure_reported_length_remaining,
which will throw the same exception as would be thrown if the
bytes up to that offset were dissected. (It does not throw an
exception if the offset is right at the end of the buffer.)
This prevents overflow and also causes invalid BER to produce
an error.

OSS-Fuzz Issue 456050170