epan: Use unsigned offsets in tvb_ensure_*_length_remaining
tvb_ensure_captured_length_remaining and
tvb_ensure_reported_length_remaining are not called with negative offsets
intended to mean "offset before the end of the (captured) length" in any
checked in code in the repository and never have been. It doesn't seem
useful to do that, particularly not for the reported length. Negative
offsets, where supported, are always counted from the end of the *captured*
data, which is not particularly well-documented and likely to be unexpected
when using a routine that measures the reported length and a tvbuff with
captured length less than the reported length.

Just require the offsets to be unsigned. This means that if a signed
integer representing an offset has overflowed, it will be treated as
a really big integer and almost surely throw an exception. That is
the desired behavior when calling these functions.

Also, document the different behavior of the two functions where the
offset is just past the end of the data. Presumably for historical
reasons, tvb_ensure_captured_length_remaining never returns 0, always
requiring at least 1 octet of data remaining, but by contrast
tvb_ensure_reported_length_remaining will return 0 for an offset just
past the end of the data.

Ping #20103