wireshark-pidl: Use tvb_new_subset_length
tvb_new_subset_length_caplen, in the case of a tvb where the parent
tvb does not contain the full requested capture length, will immediately
throw an exception when validating the resultant end offset.
We don't generally want that; if the captured length is shorter than
the reported length because a frame was truncated at capture time,
we want to take the subset of the appropriate length and only throw
a BoundsError when trying to retrieve data. That's particularly
true when there's partial data and we could add several fields that
are present. It's best to dissect as much as possible before the
exception.

Use tvb_new_subset_length() with the requested length as reported_length
instead of tvb_new_subset_length_caplen with the requested length as
the captured length and -1 as reported length.

Ping #20103, as this gets rid of a use of -1 for a length argument.

Most uses of tvb_new_subset_length_caplen, especially those with -1
for reported length, should be replaced similarly.