ISUP: Use proto_tree_add_item_ret_uint
Register multiple hf with the same filter abbreviation to cover the
different national variants for message type, which allows using
the standard proto_tree_add_item. This makes the field as a custom
column able to switch between the value and the value string entry,
and also allows filtering based on the value string entry.