SUA: Fix trivial signed overflow
The literal 256 is an int. When multiplying it by a uint16_t, the
unsigned short is promoted to an int by the integer promotions as
part of the usual arithmetic conversions (because all values of
a uint16_t can be represented by an int.) This eventually can
overflow, because what is wanted here is to combine two uint16_t
into a uint32_t.

Just make them uint32_t to start with to prevent signed overflow.
Also go ahead and use proto_tree_add_item_ret_uint.

OSS-Fuzz 458100286