brdwlk: Fix a field size (and signed overflow)
The original packet length field cannot be four bytes long (with
no bitmask), because that runs into the already dissected error
and eof fields in the next two octets. It also only contains a
value between 12 and 528. Shorten it to 2 bytes, so it cannot
overflow when multiplied by 4 to convert from words to bytes.

Use the original packet length in the case of truncation as a
reported length for the subset passed to FC.

Also, in FC, fix a case of signed overflow caused by the least
expected integer promotion rule in C - the one that converts both
operands of bitwise shift, such that an unsigned LHS smaller than
an int is converted to a [signed] int.

OSS-Fuzz 458867039