More tvb_new_subset_length_caplen conversion
More cases where the passed in captured length is the same as what
would be automatically calculated, and some cases where a reported
length from the packet is being incorrectly passed in as the captured
length instead of the reported length. If it's a truncated captured,
that causes an immediate exception instead of an exception being
thrown later when trying to retrieve data from the first missing
offset.