Abstract decode_dcerpc_add_show_list() for UI
Add an additional member to the decode_as_t struct that handles the "specialized" handling of building a Decode As list that DCE/RPC needs.  This allows the abstraction of the decode_dcerpc_add_show_list() function so that the UI is not explicitly linked to the DCE/RPC dissector.