Qt: ExportObjectDialog does not need to be Application Modal
The other statistics taps dialogs are not application modal either.
Making it be application modal makes it difficult to impossible to
use the dialog to examine frames where objects are exported in the
main window after the tapping is done and the dialog is filled out.

The same issue with the main window being frozen while tapping is
going on, and with new statistics dialogs being blocked from nested
retapping occurs as with other statistics dialogs.

Ping #20840