BT L2CAP: Fix reassembly of I-Frames
For the first fragment, [try to] copy only what is in the fragment,
not the entire reassembled SDU length, which will fail. (It will
fail on only the first pass, creating a mysterious expert info that
doesn't show up in the Expert Info dialog.)

Fixes reassembly of the I-frame SDUs beginning in frames 76 and 79
on the capture file in issue #2241; they are ultimately reassembled
into one OBEX packet in frame 81 which is correctly dissected now.

Fixup 7950fc1c941b92beb23fd29f7fed90eac9ba8c89