asn2wrs: Respect VAL_PTR for BER UTCTime
Two dissectors set the VAL_PTR directive for UTCTime, but asn2wrs.py
ignores it so the intended dissection has had no effect. Implement it.
Instead of returning the tvb length, return the string, since the
dissectors want the string.

Instead of not adding the string to the tree if the string is retrieved,
only don't add the string if the passed in HF_INDEX is -1, so that
dissectors can add the string to the tree normally but also retrieve
the string for, e.g., adding to the INFO column.

Use the default body with directives for X.509, replacing the use
of tvblen with saving the old offset.


(cherry picked from commit 461566a97aa23973854956923b1ac6e64ac1c10d)

Co-authored-by: John Thacker <johnthacker@gmail.com>