x509af, x509sat: Add UTCTime as a FT_ABSOLUTE_TIME
The ASN.1 UTCTime is a two-digit year format, which is why
dissect_ber_UTCTime treats it as a FT_STRING by default. The X.500
series define it to unambiguously mean the period 1950-2049,
with the 4-digit year GeneralizedTime to be used after that.
(Other standards are different, e.g. the X.400 standards use a
sliding window. https://www.itu.int/itudoc/itu-t/com7/solution_ww7.doc
discusses the two approaches.)

Use that to add UTCTime as a FT_ABSOLUTE_TIME in the X.500 series,
which allows natural comparisons. Switch dissect_ber_UTCTime to
use a wmem_strbuf instead of snprintf. Make the output string
ISO 8601 compliant with regards to the time zone (which it already
is in the packet.)

Ping #20848