epan: Use a GSequence for tvb_composite and make it non-recursive
To speed up tvb_composites consisting of many child tvbuffs,
use a GSequence. This is a data structure that internally has
a balanced binary tree. It differs from GTree in that one can
do initial insertions without sorting. In our case, the ordering
at the time of tvb_composite_finalize is *the* ordering, and is
the same ordering that comparing according to the calculated
end offsets will have, so this is faster on insert than a GTree.

Like a GTree, it is much faster O(log N) vs O(N) on lookup to find
the location of the TVB with end offset greater than or equal to
the desired offset than using a GQueue.

Also, make composite_memcpy iterative instead of recursive, which
removes the current hard limit of 500 without causing stack overflow
with very large numbers of child TVBs.

The file in #20860 works fine now but there could still be limits
imposed (at least on the total number of bytes?) for a compression
bomb more similar to the file found in #12077.