HTTP3: Fix reporting error codes
NGHTTP3 error codes are negative. Either store them in a signed integer
(a int16_t is large enough for all actually used values) or, if storing
them in an unsigned integer, take the unary negative when storing and
then using them. It seems easier to store them in a signed integer.

Ping #20860