ELF: Handle fuzzed data better, avoid undefined behavior
The ELF header has several overdetermined parameters, and the kernel
loader and various tools have to choose what to respect when they
are inconsistent.

In practice, the header size, program header entry size, and section
header entry size each have only two choices, reflecting the 32-bit
and 64-bit structures. The dissector currently does some calculations
(including what bytes to highlight) based on the nominal values in
the header, but advances the offsets and dissects in general using
the value corresponding to the class (32-bit or 64-bit) byte. (Note
that the kernel loader ignores the file class and data encoding byte,
apparently deducing them from the machine type, as the machine type
must be valid anyway for the program to execute.)

Consistently use the values from file class and data encoding.
If they are invalid or unknown values, add expert infos and use
the native size and endianness of the running program. Use
proto_tree_add_item_ret_uint in some places. Widen some variables
to be uint32_t as a result; this also means that they don't have
to be cast when multiplied (to prevent promotion to signed ints
instead of unsigned.)

Add a comment about how the idle detection in new_field_info doesn't
work for ELF, because the section header table is located at the
end of the file but must be read before the sections it describes,
which occur earlier.

OSS-Fuzz 459763653