Move funnel ops outside of being a tap
The "funnel menus" were being created via a tap as a convenient way to get a callback after epan was initialized. But because they are really just the backend of UI functionality, the UI should have more control over when the menus are created (and a tap callback may not be the best option).
Don't use a tap and call the funnel menu initialization directly within the UI.
Since the ops structure is a global within the funnel functionality anyway, have wslua get the value when needed, instead of caching it (and having to worry about order of initialization)