KDSP: Prevent some bogus expert info warnings
The server version and hostname fields are 32 octet
fixed length fields that contain null terminated strings,
and the bytes after the null terminators can have anything.
Change them to FT_STRINGZTRUNC, so that they don't warn
about garbage octets.

The NULL command needs to have no-op handling instead of
the default in the command switch statement to avoid a
warning about an unknown command type.

Tested with the sample capture on the SampleCaptures wiki page.