DOCSIS: Get rid of length_caplen, use maybefcs Eth dissector
Get rid of length_caplen in the DOCSIS dissector. Check to make
sure a length isn't too small to fit a fragment FCS, in which case
it would become negative after the FCS is subtracted from the length.

The DOCSIS examples on various closed issues all have the FCS included
with Ethernet frames, and the DOCSIS specifications agree that
"Normally, the Packet PDU MUST be passed across the network in its entirety,
including its original CRC," and that the PDU length field is explicitly
indicated to contain the original Ethernet CRC. (There is one attached
capture that has bogus CRCs and a missing CRC in the case of fragments,
though.) So do not use the "withoutfcs" Ethernet dissector.

For the fragment FCS, don't try to get the FCS if there aren't enough
captured bytes, because that will throw an exception too early if there
are *some* bytes, but not enough to check the FCS. Don't add the
"bad FCS" expert info twice to the tree; proto_tree_add_checksum does
that already.