IPSec TCP: Remove length_caplen
Replace tvb_new_subset_length_caplen with tvb_new_subset_length
as it should be.

Remove an "if (tree)" statement that could cause the dissector to
change whether it threw an exception or not and went on to dissect
the encapsulated payload depending on if tree was NULL.

Dissect the encapsulated payload before the trailer, so as to
throw an exception from a capture with truncated capture length
only after dissecting as many bytes as possible.