wsutil: Fix a few sign conversion issues
Fix a few internal sign conversions. Note that the gnutls_datum_t
struct has an unsigned size member, and various functions that take
a bag index take an unsigned (the function that retrieves the
count can return negative on error; check for that.)

Also check for vsnprintf() failing in one location.