fc00/cjdns: Remove variant Base32 algorithm, do not decode key
Cjdns/fc00 uses a custom Base 32 variant with a different alphabet
than that specified in RFC 4648. It's not appropriate to offer it
as "the" base32 implementation in libwsutil. (An API that offered
several implementations and the ability for a custom alphabet might
be useful.) Remove it from libwsutil.

The implementation submitted is lifted verbatim from the Cjdns
codebase without attribution, and changing the original license
from GPLv3 to GPLv2. It had never been licensed as anything other
than GPLv3 according to the git history of that file in cjdns.
The Wireshark contributor is an author on the fc00 project, but
is not clear that the proper rights were obtained to relicense
the code which originated in cjdns. Remove it, and add the key
in encoded form instead of decoding. If someone cares, they can
offer a clean room implementation of the algorithm.

Update the URL from a mirror of cjdns to the main GitHub repository.

Fix #20754


(cherry picked from commit 70a4094ae414d36813795820a31fc83666d959df)

Co-authored-by: John Thacker <johnthacker@gmail.com>