kerberos: Fix adding PA-SPAKE choice
The PA-SPAKE choice is not the value of PA-DATA, and should not be
placed in the private data struct for that. These are choices from
0 to 4; PA-SPAKE itself is PA-DATA value 151.

Just get the choice. Note that even though a ASN.1 CHOICE is always
mapped to a FT_UINT field, the "branch taken" is a signed integer,
so that -1 can be used for "no branch found." (This is less than ideal.)

Also, update the URL and comment because SPAKE is no longer an I-D
but is published as RFC 9588: https://datatracker.ietf.org/doc/rfc9588/