asterix: enum pref variable signedness
It is implementation defined whether a C enum is signed or unsigned.
(C23 adds a way to specify the underlying type.) The variables that
we register with preferences to hold the enum value use ints (not
the individual enum types), which is fine and well-defined for anything
where the values fit in an int. It's better to avoid implicitly
or explicitly changing the sign when using a pointer.

Also note that packet-asterisk uses int variables (and takes references
to them), so the struct dialog_cat_struct should also use int instead
of unsigned.