Work around a Lua 5.4.5 API botch.
Lua 5.4.5 broke the Lua API by adding a second argument to
lua_resetthread().

Fedora, and possibly other Linux distributions, compounded that error by
backporting a fix from 5.4.5 that iincluded that API-breaking change,
but didn't change LUA_VERSION_RELEASE_NUM, so that checking for
LUA_VERSION_RELEASE_NUM == 50404 isn't sufficient to detect the broken
version of the API.

See the thread starting at

https://lists.wireshark.org/archives/wireshark-dev/202511/msg00031.html

So we do it the hard way, by checking whether a program that calls the
one-argument version compiles.