Enforce -Wpointer-sign, and thus -Wall -Wextra, everywhere
The ASN.1 dissectors were the last to be converted, and -Wpointer-sign
was the last of the GCC and Clang -Wall -Wextra exceptions. Get rid
of the exception.

Fix #20175

We might still want to follow up by converting some types used by
some of the API functions, or providing extra functions for PER
and BER VAL_PTRs of different signedness, to avoid explicit casts,
which can be error-prone in the face of typos.