TPKT: Call it continuation if the length is bogus
dissect_tpkt_encap assumes that we have Continuation Data if
the first byte is not the valid version number (3). That's a
fairly weak heuristic, so if the length is illegal (i.e., < 4,
since the length includes the 4 octet TPKT header), also call it
Continuation Data for the same reason.

See #20884