BER: Fix and document dissect_ber_choice VAL_PTR usage
dissect_ber_choice returns as the VAL_PTR variable an int32_t
containing the branch_taken, not a uint32_t containing the value,
as dissect_per_choice and dissect_oer_choice (and the Integer
types) do. This is useful for detecting when no alternative
was selected.

Document this, and fix the usage and related pointer-sign warning
in the X.519 DISP dissector. Also remove a couple of obsolete
comment lines in the BER dissector header referring to "old"
versions of some of the types.

Ping #20175