PER: Fix CHOICE VAL_PTR pointer-sign warnings, document
Unlike the BER dissector, dissect_per_choice returns the tag number
(which is non-negative, see X.680 8.2) instead of the branch taken.
Like the BER dissector, the value returned is signed, not unsigned,
in order to allow -1 to be used for when the CHOICE is not present.
That may not be a sign of bad data, as the PER dissector, like the
BER dissector, does not have a way to distinguish between OPTIONAL
CHOICEs and ones that are not optional.

Fix the sign of the value pointer to be signed, and check that it
is not -1 before assuming that the choice is present. (It is unlikely
that a tag number greater than INT32_MAX is present, though somewhat
more likely than the branch number used in BER. If somehow a ChoiceType
has a tag number of UINT32_MAX, which maps to -1, this is ambiguous.)

Ping #20175